[wire-] refactored a bit unusuded fiels
