+++ config/config.yaml
@@ -1,7 +1,6 @@
 # Logging settings
-log_level: debug # NOTE: deprecated
 logging:
-  level: debug # Could be debug, info, warn, error
+  level: info # Could be debug, info, warn, error
   format: console # Could be json or console
   output: stdout # Could be stdout, stderr, or a file path like /var/log/app.log
