--- internal/common/constants.go
@@ -1,9 +1,0 @@
-package common
-
-const (
-	C_RESET  = "\033[0m"
-	C_RED    = "\033[0;31m"
-	C_GREEN  = "\033[0;32m"
-	C_BLUE   = "\033[0;34m"
-	C_YELLOW = "\033[1;33m"
-)
